Moved focus methods up in the class heirarchy
